added video support via ffmpeg and libav to animate_move()

+++ R/get_imconvert.R
@@ -1,15 +1,18 @@
 #' Search/download and install ImageMagick
 #'
 #' \code{get_imconvert} searches for the convert tool being part of the ImageMagick software on your system (either Windows or Linux) or installs it.
-#' The convert tool system command or directory is needed by the \code{animate_move()} function of the \code{moveVis} package.
-#' If you are running Windows, the function can download and either temporarily or permanently install ImageMagick
+#' The convert tool system command or directory is needed by the \code{animate_move()} function to support the \code{.gif} output format.
+#'  
+#' @param dir character. Directory were to download, unzip and install ImageMagick. If set to "auto", a temporary directory is used. Default is "auto".
+#' @param ... additional arguments. Currently not used.
+#' 
+#' @return The convert tool command line command or directory. The output can serve as \code{conv_dir} input to the \code{animate_move} function.
+#' 
+#' @details If you are running Windows, the function can download and either temporarily or permanently install ImageMagick
 #' on your Windows system in case that no existing installation can be found. If you are running Linux,
 #' the function provides a root permission requiring command to be executed once by the user in the terminal
 #' to install ImageMagick (Linux), if ImageMagick is not installed. On standard Ubuntu distributions, ImageMagick belongs to the preinstalled packages by default.
-#' 
-#' @param dir character. Directory were to download, unzip and install ImageMagick. If set to "auto", a temporary directory is used. Default is "auto".
-#' 
-#' @return The convert tool command line command or directory. The output can serve as \code{conv_dir} input to the \code{animate_move} function.
+#' If you are running macOS (former OSX), please install ImageMagick manually from \url{https://www.imagemagick.org/script/download.php} or via a package manager such as 'brew'. 
 #' 
 #' @examples
 #' conv_dir <- get_imconvert()
@@ -21,40 +24,50 @@
 #' @importFrom RCurl getURL
 #' @export
 
-get_imconvert <- function(dir = "auto"){
+get_imconvert <- function(dir = "auto", ...){
+  
+  arg <- list(...)
+  s_try <- try(arg$nodownload)
+  if(class(s_try) == "NULL"){nodownload <-  FALSE}else{nodownload <- arg$nodownload}
   
   #Define output handling
   log_level <- 1
-  out <- function(input,type = 1){
-    signs <- c("[LOG]: ", "[WARNING]: ")
-    if(type == 2 & log_level <= 2){warning(paste(signs[2],input))}
-    else{if(type == 3){stop(input,call. = FALSE)}else{if(log_level == 1){cat(paste(signs[1],input),sep="\n")}}}
+  out <- function(input,type = 1, ll = log_level, msg = FALSE){
+    signs <- c("", "")
+    if(type == 2 & ll <= 2){warning(paste0(signs[2],input), call. = FALSE, immediate. = TRUE)}
+    else{if(type == 3){stop(input,call. = FALSE)}else{if(ll == 1){
+      if(msg == FALSE){cat(paste0(signs[1],input),sep="\n")
+      }else{message(paste0(signs[1],input))}}}}
   }
   
+  if(.Platform$OS.type == 'windows'){cmd.fun <- shell}else{cmd.fun <- system}
   if(dir == "auto"){dir <- tempdir()}
-  if(.Platform$OS.type == 'windows'){
-    if(length(grep("convert.exe",list.files(paste0("C:/Program Files/",grep("ImageMagick", list.files("C:/Program Files/"),value = TRUE))))) != 0){
-      conv_dir <- paste0("C:/Program Files/", list.files("C:/Program Files/")[grep("ImageMagick",list.files("C:/Program Files/"))], "/convert.exe")
-    }else{
-      if(!file.exists(paste0(dir,"/imagick/convert.exe"))){
-        print("Downloading portable ImageMagick...")
-        ftp.dir <- "ftp://ftp.imagemagick.org/pub/ImageMagick/binaries/"
-        zip.dir <- grep(".zip$",grep('portable', unlist(strsplit(getURL(ftp.dir,dirlistonly = TRUE),"[\\\\]|[^[:print:]]",fixed=FALSE)), value=TRUE),value = TRUE)
-        f.dir <- grep(unlist(strsplit(Sys.getenv("R_ARCH"),"/"))[2],zip.dir,value = TRUE)
-        f.dir <- paste0(ftp.dir,f.dir[length(f.dir)])
-        download.file(url = f.dir, destfile = paste0(dir,"/imagick.zip"),method="auto")
-        unzip(paste0(dir,"/imagick.zip"),exdir = paste0(dir,"/imagick"))
-        file.remove(paste0(dir,"/imagick.zip"))
+  
+  tryit <- try(cmd.fun("convert",ignore.stdout = TRUE,ignore.stderr = TRUE))
+  if(tryit != 1){
+    if(.Platform$OS.type == 'windows'){
+      if(length(grep("convert.exe",list.files(paste0("C:/Program Files/",grep("ImageMagick", list.files("C:/Program Files/"),value = TRUE))))) != 0){
+        conv_dir <- paste0("C:/Program Files/", list.files("C:/Program Files/")[grep("ImageMagick",list.files("C:/Program Files/"))], "/convert.exe")
+      }else{
+        if(!file.exists(paste0(dir,"/imagick/convert.exe"))){
+          if(nodownload == TRUE){out("'convert' could not be located on this system. Please use get_imconvert() to install ImageMagick or install manually from 'https://www.imagemagick.org/script/download.php'",type=3)
+          }else{
+            print("Downloading portable ImageMagick...")
+            ftp.dir <- "ftp://ftp.imagemagick.org/pub/ImageMagick/binaries/"
+            zip.dir <- grep(".zip$",grep('portable', unlist(strsplit(getURL(ftp.dir,dirlistonly = TRUE),"[\\\\]|[^[:print:]]",fixed=FALSE)), value=TRUE),value = TRUE)
+            f.dir <- grep(unlist(strsplit(Sys.getenv("R_ARCH"),"/"))[2],zip.dir,value = TRUE)
+            f.dir <- paste0(ftp.dir,f.dir[length(f.dir)])
+            download.file(url = f.dir, destfile = paste0(dir,"/imagick.zip"),method="auto")
+            unzip(paste0(dir,"/imagick.zip"),exdir = paste0(dir,"/imagick"))
+            file.remove(paste0(dir,"/imagick.zip"))
+          } 
+        }
+        conv_dir <- paste0(dir,"\\imagick\\convert.exe")
       }
-      conv_dir <- paste0(dir,"\\imagick\\convert.exe")
+      return(conv_dir)
+    }else{
+      out("No ImageMagick installation could be found. Please install manually. On Linux, open the terminal, enter 'sudo apt-get install imagemagick'.
+          On other systems, install manually from 'https://www.imagemagick.org/script/download.php'.",type=3)
     }
-    return(conv_dir)
-  }else{
-    tryit <- try(system("convert",ignore.stdout = TRUE,ignore.stderr = TRUE))
-    if(tryit != 1){
-      out("No ImageMagick installation could be found. Please install manually.",type=1)
-      out("On Linux, open the terminal, enter 'sudo apt-get install imagemagick'.",type=1)
-      out("On other systems, install manually from 'https://www.imagemagick.org/script/download.php'.",type=1)
-    }else{conv_dir <- "convert"; return(conv_dir)}
-  }
+  }else{conv_dir <- "convert"; return(conv_dir)}
 }
